requirements.txt
pandas
numpy
scikit-learn
matplotlib
tqdm
python-dotenv
jupyter
openai
